pyproject.toml
[tool.poetry]
name = "autobot-tg"
version = "0.1.2a2"
description = "Telegram Bot creation made easy"
authors = ["Andrew S. <andrewsapw@gmail.com>"]
license = "BSD-3-Clause"
readme = "README.md"
homepage = "https://github.com/andrewsapw/autobot"
repository = "https://github.com/andrewsapw/autobot"
include = [
    "LICENSE",
]
packages = [
    { include = "autobot" },
]

[tool.poetry.scripts]
autobot = "autobot.__main__:main"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^1.10.2"
tomli = "^2.0.1"
uvloop = "^0.17.0"
pyyaml = "^6.0"
aiogram = {version = "^3.0.0b5", allow-prereleases = true}
python-dotenv = "^0.21.0"
loguru = "^0.6.0"
networkx = "^2.8.8"
attrs = "^22.1.0"
jinja2 = "^3.1.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
black = {version = "^22.10.0", allow-prereleases = true}
autoflake = "^1.7.7"
pre-commit = "^2.20.0"
coverage = "^6.5.0"
pynvim = "^0.4.3"
ipython = "^8.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
